+++ src/repositories/sessionRepository.js
@@ -1,4 +1,4 @@
-import {SqlRepository} from "./Repository.js";
+import {SqlRepository} from "./_sqlRepository.js";
 
 export class SessionRepository extends SqlRepository {
     get insertIntoSql() {
